Classical Basis Layers

The following basis are helper functions for easily building arrays of the form [f0(x), ..., f{n-1}(x)], where f is the corresponding function of the basis (e.g, Chebyshev Polynomials, Legendre Polynomials, etc.)

DiffEqFlux.ChebyshevBasis — Type

Constructs a Chebyshev basis of the form [T{0}(x), T{1}(x), ..., T{n-1}(x)] where Tj(.) is the j-th Chebyshev polynomial of the first kind.

ChebyshevBasis(n)

Arguments:

n: number of terms in the polynomial expansion.

DiffEqFlux.SinBasis — Type

Constructs a sine basis of the form [sin(x), sin(2x), ..., sin(nx)].

SinBasis(n)

Arguments:

n: number of terms in the sine expansion.

DiffEqFlux.CosBasis — Type

Constructs a cosine basis of the form [cos(x), cos(2x), ..., cos(nx)].

CosBasis(n)

Arguments:

n: number of terms in the cosine expansion.

DiffEqFlux.FourierBasis — Type

Constructs a Fourier basis of the form Fj(x) = j is even ? cos((j÷2)x) : sin((j÷2)x) => [F0(x), F1(x), ..., Fn(x)].

FourierBasis(n)

Arguments:

n: number of terms in the Fourier expansion.

DiffEqFlux.LegendreBasis — Type

Constructs a Legendre basis of the form [P{0}(x), P{1}(x), ..., P{n-1}(x)] where Pj(.) is the j-th Legendre polynomial.

LegendreBasis(n)

Arguments:

n: number of terms in the polynomial expansion.

DiffEqFlux.PolynomialBasis — Type

Constructs a Polynomial basis of the form [1, x, ..., x^(n-1)].

PolynomialBasis(n)

Arguments:

n: number of terms in the polynomial expansion.