Classical Basis Layers

The following basis are helper functions for easily building arrays of the form [f_0(x), ..., f_{n-1}(x)], where f is the corresponding function of the basis (e.g, Chebyshev Polynomials, Legendre Polynomials, etc.)

DiffEqFlux.ChebyshevBasis — Function

ChebyshevBasis(n)

Constructs a Chebyshev basis of the form [T_{0}(x), T_{1}(x), ..., T_{n-1}(x)] where T_j(.) is the j-th Chebyshev polynomial of the first kind.

Arguments:

n: number of terms in the polynomial expansion.

DiffEqFlux.SinBasis — Function

SinBasis(n)

Constructs a sine basis of the form [sin(x), sin(2x), ..., sin(nx)].

Arguments:

n: number of terms in the sine expansion.

DiffEqFlux.CosBasis — Function

CosBasis(n)

Constructs a cosine basis of the form [cos(x), cos(2x), ..., cos(nx)].

Arguments:

n: number of terms in the cosine expansion.

DiffEqFlux.FourierBasis — Function

FourierBasis(n)

Constructs a Fourier basis of the form F_j(x) = j is even ? cos((j÷2)x) : sin((j÷2)x) => [F_0(x), F_1(x), ..., F_n(x)].

Arguments:

n: number of terms in the Fourier expansion.

DiffEqFlux.LegendreBasis — Function

LegendreBasis(n)

Constructs a Legendre basis of the form [P_{0}(x), P_{1}(x), ..., P_{n-1}(x)] where P_j(.) is the j-th Legendre polynomial.

Arguments:

n: number of terms in the polynomial expansion.

DiffEqFlux.PolynomialBasis — Function

PolynomialBasis(n)

Constructs a Polynomial basis of the form [1, x, ..., x^(n-1)].

Arguments:

n: number of terms in the polynomial expansion.