Hamiltonian Neural Network

The following layer helps construct a neural network which allows learning dynamics and conservation laws by approximating the hamiltonian of a system.

DiffEqFlux.HamiltonianNN — Type

HamiltonianNN(model; ad = AutoForwardDiff())

Constructs a Hamiltonian Neural Network [1]. This neural network is useful for learning symmetries and conservation laws by supervision on the gradients of the trajectories. It takes as input a concatenated vector of length 2n containing the position (of size n) and momentum (of size n) of the particles. It then returns the time derivatives for position and momentum.

Note

This doesn't solve the Hamiltonian Problem. Use NeuralHamiltonianDE for such applications.

Arguments:

model: A Flux.Chain or Lux.AbstractExplicitLayer neural network that returns the Hamiltonian of the system.

ad: The autodiff framework to be used for the internal Hamiltonian computation. The default is AutoForwardDiff()

Note

If training with Zygote, ensure that the chunksize for AutoForwardDiff is set to nothing.

References:

[1] Greydanus, Samuel, Misko Dzamba, and Jason Yosinski. "Hamiltonian Neural Networks." Advances in Neural Information Processing Systems 32 (2019): 15379-15389.

DiffEqFlux.NeuralHamiltonianDE — Type

NeuralHamiltonianDE(model, tspan, args...; kwargs...)

Contructs a Neural Hamiltonian DE Layer for solving Hamiltonian Problems parameterized by a Neural Network HamiltonianNN.

Arguments:

model: A Flux.Chain, Lux.AbstractExplicitLayer, or Hamiltonian Neural Network that predicts the Hamiltonian of the system.

tspan: The timespan to be solved on.

kwargs: Additional arguments splatted to the ODE solver. See the Common Solver Arguments documentation for more details.