Neural Differential Equation Layer Functions

The following layers are helper functions for easily building neural differential equation architectures in the currently most efficient way. As demonstrated in the tutorials, they do not have to be used since automatic differentiation will just work over solve, but these cover common use cases and choose what's known to be the optimal mode of AD for the respective equation type.

DiffEqFlux.NeuralODE — Type

Constructs a continuous-time recurrant neural network, also known as a neural ordinary differential equation (neural ODE), with a fast gradient calculation via adjoints [1]. At a high level this corresponds to solving the forward differential equation, using a second differential equation that propagates the derivatives of the loss backwards in time.

NeuralODE(model,tspan,alg=nothing,args...;kwargs...)

Arguments:

model: A Flux.Chain or Lux.AbstractExplicitLayer neural network that defines the ̇x.

tspan: The timespan to be solved on.

alg: The algorithm used to solve the ODE. Defaults to nothing, i.e. the default algorithm from DifferentialEquations.jl.

sensealg: The choice of differentiation algorthm used in the backpropogation. Defaults to an adjoint method. See the Local Sensitivity Analysis documentation for more details.

kwargs: Additional arguments splatted to the ODE solver. See the Common Solver Arguments documentation for more details.

References:

[1] Pontryagin, Lev Semenovich. Mathematical theory of optimal processes. CRC press, 1987.

DiffEqFlux.NeuralDSDE — Type

Constructs a neural stochastic differential equation (neural SDE) with diagonal noise.

NeuralDSDE(drift,diffusion,tspan,alg=nothing,args...;
           sensealg=TrackerAdjoint(),kwargs...)

Arguments:

drift: A Flux.Chain or Lux.AbstractExplicitLayer neural network that defines the drift function.

diffusion: A Flux.Chain or Lux.AbstractExplicitLayer neural network that defines the diffusion function. Should output a vector of the same size as the input.

tspan: The timespan to be solved on.

alg: The algorithm used to solve the ODE. Defaults to nothing, i.e. the default algorithm from DifferentialEquations.jl.

sensealg: The choice of differentiation algorthm used in the backpropogation. Defaults to using reverse-mode automatic differentiation via Tracker.jl

kwargs: Additional arguments splatted to the ODE solver. See the Common Solver Arguments documentation for more details.

DiffEqFlux.NeuralSDE — Type

Constructs a neural stochastic differential equation (neural SDE).

NeuralSDE(drift,diffusion,tspan,nbrown,alg=nothing,args...;
          sensealg=TrackerAdjoint(),kwargs...)

Arguments:

drift: A Flux.Chain or Lux.AbstractExplicitLayer neural network that defines the drift function.

diffusion: A Flux.Chain or Lux.AbstractExplicitLayer neural network that defines the diffusion function. Should output a matrix that is nbrown x size(x,1).

tspan: The timespan to be solved on.

nbrown: The number of Brownian processes

alg: The algorithm used to solve the ODE. Defaults to nothing, i.e. the default algorithm from DifferentialEquations.jl.

sensealg: The choice of differentiation algorthm used in the backpropogation. Defaults to using reverse-mode automatic differentiation via Tracker.jl

kwargs: Additional arguments splatted to the ODE solver. See the Common Solver Arguments documentation for more details.

Missing docstring.

Missing docstring for NeuralCDDE. Check Documenter's build log for details.

DiffEqFlux.NeuralDAE — Type

Constructs a neural differential-algebraic equation (neural DAE).

NeuralDAE(model,constraints_model,tspan,alg=nothing,args...;
          sensealg=TrackerAdjoint(),kwargs...)

Arguments:

model: A Flux.Chain or Lux.AbstractExplicitLayer neural network that defines the derivative function. Should take an input of size x and produce the residual of f(dx,x,t) for only the differential variables.

constraints_model: A function constraints_model(u,p,t) for the fixed constaints to impose on the algebraic equations.

tspan: The timespan to be solved on.

alg: The algorithm used to solve the ODE. Defaults to nothing, i.e. the default algorithm from DifferentialEquations.jl.

sensealg: The choice of differentiation algorthm used in the backpropogation. Defaults to using reverse-mode automatic differentiation via Tracker.jl

kwargs: Additional arguments splatted to the ODE solver. See the Common Solver Arguments documentation for more details.

DiffEqFlux.NeuralODEMM — Type

Constructs a physically-constrained continuous-time recurrant neural network, also known as a neural differential-algebraic equation (neural DAE), with a mass matrix and a fast gradient calculation via adjoints [1]. The mass matrix formulation is:

\[Mu' = f(u,p,t)\]

where M is semi-explicit, i.e. singular with zeros for rows corresponding to the constraint equations.

NeuralODEMM(model,constraints_model,tspan,mass_matrix,alg=nothing,args...;kwargs...)

Arguments:

model: A Flux.Chain or Lux.AbstractExplicitLayer neural network that defines the ̇f(u,p,t)

constraints_model: A function constraints_model(u,p,t) for the fixed constaints to impose on the algebraic equations.

tspan: The timespan to be solved on.

mass_matrix: The mass matrix associated with the DAE

alg: The algorithm used to solve the ODE. Defaults to nothing, i.e. the default algorithm from DifferentialEquations.jl. This method requires an implicit ODE solver compatible with singular mass matrices. Consult the DAE solvers documentation for more details.

sensealg: The choice of differentiation algorthm used in the backpropogation. Defaults to an adjoint method. See the Local Sensitivity Analysis documentation for more details.

kwargs: Additional arguments splatted to the ODE solver. See the Common Solver Arguments documentation for more details.

Missing docstring.

Missing docstring for AugmentedNDELayer. Check Documenter's build log for details.

Helper Layer Functions

DiffEqFlux.DimMover — Type

Constructs a Dimension Mover Layer.

DimMover(from, to)

DiffEqFlux.FluxBatchOrder — Function

We can have Flux's conventional order (data, channel, batch) by using it as the last layer of Flux.Chain to swap the batch-index and the time-index of the Neural DE's output. considering that each time point is a channel.

FluxBatchOrder = DimMover(-2, -1)