Tensor Product Layer

The following layer is a helper function for easily constructing a TensorLayer, which takes as input an array of n tensor product basis, $[B_1, B_2, ..., B_n]$, a data point x, computes $z[i] = W[i,:] ⨀ [B_1(x[1]) ⨂ B_2(x[2]) ⨂ ... ⨂ B_n(x[n])]$, where W is the layer's weight, and returns [z[1], ..., z[out]].

DiffEqFlux.TensorLayer — Type

Constructs the Tensor Product Layer, which takes as input an array of n tensor product basis, [B1, B2, ..., Bn] a data point x, computes z[i] = W[i,:] ⨀ [B1(x[1]) ⨂ B2(x[2]) ⨂ ... ⨂ Bn(x[n])], where W is the layer's weight, and returns [z[1], ..., z[out]].

TensorLayer(model,out,p=nothing)

Arguments:

model: Array of TensorProductBasis [B1(n1), ..., Bk(nk)], where k corresponds to the dimension of the input.

out: Dimension of the output.

p: Optional initialization of the layer's weight. Initizalized to 0 by default.